move lineage SPI models to extensions module
