+++ extensions/lineage/src/main/java/org/apache/polaris/extensions/lineage/LineageData.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.lineage;
+package org.apache.polaris.extensions.lineage;
 
 import java.util.Objects;
 import java.util.OptionalLong;
