+++ extensions/lineage/src/main/java/org/apache/polaris/extensions/lineage/LineageNodeType.java
@@ -16,10 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.lineage;
+package org.apache.polaris.extensions.lineage;
 
-/** Supported query granularities for lineage lookups. */
-public enum LineageGranularity {
+/** Node kinds surfaced by normalized lineage queries. */
+public enum LineageNodeType {
   DATASET,
   COLUMN
 }
