+++ extensions/lineage/src/main/java/org/apache/polaris/extensions/lineage/LineageQueryRequest.java
@@ -16,26 +16,16 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.lineage;
+package org.apache.polaris.extensions.lineage;
 
-import java.util.List;
 import java.util.Objects;
 
-/** A node returned in a lineage query response. */
-public record LineageNode(
-    String id,
-    LineageNodeType type,
-    LineageData data,
-    boolean opaque,
-    List<LineageFieldMapping> fieldMappings) {
-  public LineageNode {
-    Objects.requireNonNull(id, "id must be non-null");
-    Objects.requireNonNull(type, "type must be non-null");
-    fieldMappings =
-        List.copyOf(Objects.requireNonNull(fieldMappings, "fieldMappings must be non-null"));
-  }
-
-  public LineageNode(String id, LineageNodeType type, LineageData data, boolean opaque) {
-    this(id, type, data, opaque, List.of());
+/** Request model for normalized lineage lookups. */
+public record LineageQueryRequest(
+    String nodeId, LineageDirection direction, LineageGranularity granularity) {
+  public LineageQueryRequest {
+    Objects.requireNonNull(nodeId, "nodeId must be non-null");
+    Objects.requireNonNull(direction, "direction must be non-null");
+    Objects.requireNonNull(granularity, "granularity must be non-null");
   }
 }
