+++ gradle/projects.main.properties
@@ -20,7 +20,7 @@
 
 polaris-bom=bom
 polaris-core=polaris-core
-polaris-lineage=lineage
+polaris-extensions-lineage=extensions/lineage
 polaris-api-iceberg-service=api/iceberg-service
 polaris-api-management-model=api/management-model
 polaris-api-management-service=api/management-service
