+++ runtime/service/build.gradle.kts
@@ -26,7 +26,7 @@ plugins {
 
 dependencies {
   implementation(project(":polaris-core"))
-  implementation(project(":polaris-lineage"))
+  implementation(project(":polaris-extensions-lineage"))
   implementation(project(":polaris-api-management-model"))
   implementation(project(":polaris-api-management-service"))
   implementation(project(":polaris-api-iceberg-service"))
