+++ runtime/service/src/test/java/org/apache/polaris/service/lineage/DefaultLineageServiceTest.java
@@ -24,9 +24,9 @@ import static org.mockito.Mockito.when;
 import org.apache.polaris.core.config.FeatureConfiguration;
 import org.apache.polaris.core.config.RealmConfig;
 import org.apache.polaris.core.context.CallContext;
-import org.apache.polaris.lineage.LineageDirection;
-import org.apache.polaris.lineage.LineageGranularity;
-import org.apache.polaris.lineage.LineageQueryRequest;
+import org.apache.polaris.extensions.lineage.LineageDirection;
+import org.apache.polaris.extensions.lineage.LineageGranularity;
+import org.apache.polaris.extensions.lineage.LineageQueryRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
